refactor: rename transport auth module
