+++ codex-rs/app-server/src/transport/auth.rs
